chore: rename filip.types
